Create README.md
## Extended Description

Welcome to the Market Basket Analysis Project.
This project focuses on analyzing customer purchase patterns to uncover associations between different items. The primary goal is to provide insights that can be valuable for marketing and product placement strategies. This repository provides a brief overview, but for more detailed information and in-depth documentation, check out my Medium publication: https://medium.com/@chemistry8526/boosting-sales-with-data-the-power-of-market-basket-analysis-in-retail-c79cc10a14df